Vundle not installed fix
Was previously looking for a README.md file in the Vundle folder, this
failed when the Vundle plugin manager deleted it during a PluginClean
operation. I simplified the Vundle installed check to just look for the
Vundle directory and to throw an error it it is not found, telling the
user to install it.