Modified vimrc
Python initialization.
For systems that have pyenv installed and a vim with both python 2 and 3
support builtin, make sure that the pyenv active python is used first.
This allows a vim that does not support having both python versions
available to at least activate the current pyenv version for the current
vim session. Also echo a message if we are unable to initalize the
currently available python version.